requirements.txt
alembic==1.7.5
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
